Add files via upload
Hi po, sir! I made a grades database with a GWA calculator feature that updates the GWA on the tabular view as soon as you make an entry. I based it on the previous exercise and my GUI program to be ran at the virtual environment is the GCalcAppCtk.py (like the EmpAppCtk.py in the previous exercise.) My unit test is also in that file. For the methods I used for the add, entry, update, etc., they're all in GCalcDb.py (just like in EmpDb.py).